added semantic and add code to the footer and header
